% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_bootstrap.R
\name{fr_bootstrap}
\alias{fr_bootstrap}
\title{Bootstrapping the fit to get confidence intervals}
\usage{
fr_bootstrap(
  fit,
  bt_iter = 1000,
  MC = T,
  noC = ceiling(parallel::detectCores()),
  mle2_tol = 1e-06
)
}
\arguments{
\item{fit}{A mle2 object.}

\item{bt_iter}{The number of bootstrap iterations.}

\item{MC}{Use parallel computing? Actually, this is not a question!! (F/T)}

\item{noC}{If MC = T, how many cores to use? Use all ist the default.}

\item{mle2_tol}{The tolerance of a single mle2 fit.}
}
\value{
Returns a list with (1) the confidence limits, and (2) the bootstrapped parameter samples.
}
\description{
TBA
}
\details{
TBA
}
\examples{

data(dfr_vp)

fr_fit(
  Neaten = dfr_vp$Neaten,
  Nstart = dfr_vp$Nstart,
  P = rep(1, length(dfr_vp$Neaten)),
  tend = rep(1, length(dfr_vp$Neaten))
)

# only run if you want to take a break...
system.time({
 frbt <- fr_bootstrap(fit)
})
CIs:
frbt[1]

}
