% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_CI_profile.R
\name{fr_CI_profile}
\alias{fr_CI_profile}
\title{Profiling the confidence intervalls in parallel}
\usage{
fr_CI_profile(fit, MC = T, noC = 3)
}
\arguments{
\item{fit}{The prey/resource items eaten throughout the experimental
trial.}

\item{MC}{Use parallel computing? (F/T)}

\item{noC}{If MC = T, how many cores to use?}
}
\value{
Returns the profiled confidence intervals.
}
\description{

}
\examples{

data(dfr_vp)

fit <- fr_fit(
  Neaten = dfr_vp$Neaten,
  Nstart = dfr_vp$Nstart,
  P = rep(1, length(dfr_vp$Neaten)),
  tend = rep(1, length(dfr_vp$Neaten))
)

CI <- fr_confint(fit = fit)
CI

}
