% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_fit.R
\name{fr_fit}
\alias{fr_fit}
\title{Fitting of the functional response model with simulation by lsoda()}
\usage{
fr_fit(Nstart, Neaten, P, tend, tsteps = 100)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{tend}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}
}
\value{
Returns a vector of negative log-likelihoods in the order of
    Neaten/Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

data(dfr_vp)

fr_fit(
  Neaten = dfr_vp$Neaten,
  Nstart = dfr_vp$Nstart,
  P = rep(1, length(dfr_vp$Neaten)),
  tend = rep(1, length(dfr_vp$Neaten))
)

}
