% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_model.R
\name{fr_model}
\alias{fr_model}
\title{The Functional Response Model}
\usage{
fr_model(Nstart, P, tend, Fmax, N0, q, tsteps)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{P}{The consumer/predator abundance/counts/density.}

\item{tend}{The end time over up to which should be simulated (e.g. 24h).}

\item{Fmax}{The maximum feeding rate.}

\item{N0}{The half saturation density.}

\item{q}{The functional response shape parameter.}

\item{tsteps}{The number of steps that should be returned by the solver.}
}
\value{
Returns the prey/resource density at Tend.
}
\description{
TBA
}
\examples{

fr_model(
  Nstart = 25,
  Fmax = 2,
  N0 = 5,
  q = 1,
  P = 1,
  tend = 24,
  tsteps = 100
 )

}
