% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_nll.R
\name{fr_nll}
\alias{fr_nll}
\title{Calculation of negative log-likelihoods using lsoda() as method}
\usage{
fr_nll(
  Neaten,
  Nstart,
  P,
  tend,
  l10_Fmax,
  l10_N0,
  h,
  penalty = 100,
  h_low = 1,
  h_up = 5,
  tsteps = 100
)
}
\arguments{
\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{Nstart}{The initial prey/resource density.}

\item{P}{The consumer/predator abundance/counts/density.
Must be a vector of the length of Nstart.}

\item{tend}{The end time over up to which should be simulated (e.g. 24h).
Must be a vector of the length of Nstart.}

\item{l10_Fmax}{The log10 of the maximum feeding rate.}

\item{l10_N0}{The log10 of the half saturation density.}

\item{h}{The Hill exponent. It is functional response shape parameter.}

\item{tsteps}{The number of steps that should be returned by the solver.}
}
\value{
Returns a single negative log-likelihood value.
}
\description{
Calculation of negative log-likelihoods using lsoda() as method
}
\examples{

TBA


}
