% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_rll.R
\name{fr_rll}
\alias{fr_rll}
\title{Calculation of a relative log-likelihood for CI estimation}
\usage{
fr_rll(
  x,
  which_parm,
  min_nll,
  l10_Fmax_in,
  l10_N0_in,
  h_in,
  fit,
  CI_level,
  df,
  mle2_tol = 1e-08,
  for_uniroot = F
)
}
\arguments{
\item{x}{TBA}

\item{which_parm}{TBA}

\item{min_nll}{TBA}

\item{l10_Fmax_in}{The log10 of the maximum feeding rate.}

\item{l10_N0_in}{The log10 of the half saturation density.}

\item{h_in}{The Hill exponent. It is functional response shape parameter.}

\item{fit}{TBA}

\item{CI_level}{TBA}

\item{df}{TBA}

\item{mle2_tol}{TBA}

\item{for_uniroot}{TBA}
}
\value{
TBA
}
\description{

}
\examples{

TBA


}
