% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_sim.R
\name{fr_sim}
\alias{fr_sim}
\title{Calculation of the Functional Response Model using odeintr}
\usage{
fr_sim(Nstart, P, tend, Fmax, N0, q, tsteps = 100)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{P}{The consumer/predator abundance/counts/density. Must be a vector of
the length of Nstart.}

\item{tend}{The end time over up to which should be simulated (e.g. 24h).
Must be a vector of the length of Nstart.}

\item{Fmax}{The maximum feeding rate.}

\item{N0}{The half saturation density.}

\item{q}{The functional response shape parameter.}

\item{tsteps}{The number of steps that should be returned by the solver.}
}
\value{
Returns a vector with the number of prey/resource eaten in the order
    of Nstart.
}
\description{
TBA
}
\examples{

Nstart <- 1:100

y1 <- fr_sim(
  Nstart = Nstart,
  P = rep(1, length(Nstart)),
  tend = rep(24, length(Nstart)),
  Fmax = 2,
  N0 = 25,
  q = .5
)

plot(Nstart, y1, type="l")

}
