% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frfit_deSolve.R
\name{frfit_deSolve}
\alias{frfit_deSolve}
\title{Fitting of the functional response model with simulation by lsoda()}
\usage{
frfit_deSolve(Nstart, Neaten, P, tmax, tsteps = 100)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{tmax}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}
}
\value{
Returns a vector of negative log-likelihoods in the order of
    Neaten/Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

data(vucic_pestic_fr)

frfit_deSolve(
  Neaten = vucic_pestic_fr$Neaten,
  Nstart = vucic_pestic_fr$N0,
  P = 1,
  tmax = 24
)

}
