% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmodel_deSolve.R
\name{frmodel_deSolve}
\alias{frmodel_deSolve}
\title{The Functional Response Model for `deSolve`}
\usage{
frmodel_deSolve(t, x, parms)
}
\arguments{
\item{t}{The time variable as required by e.g. \link[deSolve]{lsoda}.}

\item{x}{The density from which x at t+1 should be computed, see e.g.
\link[deSolve]{lsoda}}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}
}
\value{
Returns a list of x-values and dependent y-values.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

frmodel_deSolve(
  t = 0, # only required for simulations using e.g. lsoda()
  x = 10, # the density at wich x at t+1 should be computed
  parms = list(
    Fmax = 10,
    N0 = 5,
    q = 1,
    P = 1
    )
 )

}
