% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmodel_odeintr.R
\name{frmodel_odeintr}
\alias{frmodel_odeintr}
\title{The Functional Response Model}
\usage{
frmodel_odeintr(Nstart, Fmax, N0, q, P, tend, steps)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{Fmax}{The maximum feeding rate.}

\item{N0}{The half saturation density.}

\item{q}{The functional response shape parameter.}

\item{P}{The consumer/predator abundance/counts/density.}

\item{tend}{The time of the end of the time series (e.g. 24h)}

\item{steps}{The number of steps at time > 0 at which \link[odeintr] should
compute the result.}
}
\value{
Returns the prey/resource density at Tend.
}
\description{
TBA
}
\examples{

frmodel_odeintr(
  Nstart = 10,
  Fmax = 10,
  N0 = 5,
  q = 1,
  P = 1,
  tend = 24,
  steps = 100
 )

}
