% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frnll_deSolve.R
\name{frnll_deSolve}
\alias{frnll_deSolve}
\title{Calculation of negative log-likelihoods using lsoda() as method}
\usage{
frnll_deSolve(Neaten, Nstart, Fmax, N0, q, P, tmax, tsteps = 100)
}
\arguments{
\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{Nstart}{The initial prey/resource density.}

\item{tmax}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}
}
\value{
Returns a vector of negative log-likelihoods in the order of
    Neaten/Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

Nstart <- rep(1:100,10)
Neaten <- rbinom(length(Nstart),
                 size = Nstart,
                 prob = .5
                 )

frnll_deSolve(
  Neaten = Neaten,
  Nstart = Nstart,
  Fmax = 10,
  N0 = 5,
  q = 1,
  P = 1,
  tmax = 24
)

frnll_deSolve(
  Neaten = Neaten,
  Nstart = rep(1:100,10),
  Fmax = 10,
  N0 = 5,
  q = 1,
  P = 1,
  tmax = rnorm(length(Nstart), 24, sd = 2)
)

}
