% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frsim_deSolve.R
\name{frsim_deSolve}
\alias{frsim_deSolve}
\title{Calculation of the Functional Response Model using odeintr deSolve}
\usage{
frsim_deSolve(Nstart, parms, tmax, steps = 100)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}

\item{tmax}{The time over which should be simulated.}

\item{steps}{The number of times steps that \link[deSolve]{lsoda} should
return.}
}
\value{
Returns a vector with the number of prey/resource eaten in the order
    of Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

Nstart <- rep(1:100,10)

frsim_deSolve(
  Nstart = Nstart,
  parms = list(
    Fmax = 10,
    N0 = 5,
    q = 1,
    P = 1
  ),
  tmax = 24
)

frsim_deSolve(
  Nstart = Nstart,
  parms = list(
    Fmax = 10,
    N0 = 5,
    q = 1,
    P = 1
  ),
  tmax = rnorm(length(Nstart), 24, sd = 2)
)

}
