% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs_parms.R
\name{lhs_parms}
\alias{lhs_parms}
\title{Random Sampling of Parameters and Selection based on lhs}
\usage{
lhs_parms(
  Neaten,
  Nstart,
  P,
  tend,
  l10_Fmax_range,
  l10_N0_range,
  h_range,
  tsteps = 100,
  lhs_iter = 1000,
  MC = T,
  noC = ceiling(parallel::detectCores()/2)
)
}
\arguments{
\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{Nstart}{The initial prey/resource density.}

\item{P}{The consumer/predator abundance/counts/density.
Must be a vector of the length of Nstart.}

\item{tend}{The end time over up to which should be simulated (e.g. 24h).
Must be a vector of the length of Nstart.}

\item{tsteps}{The number of steps that should be returned by the solver.}

\item{lhs_iter}{Number of random latin hypercube samplings.}

\item{MC}{Use parallel computing? (F/T)}

\item{noC}{If MC = T, how many cores to use?}

\item{l10_Fmax}{The log10 of the maximum feeding rate.}

\item{l10_N0}{The log10 of the half saturation density.}

\item{h}{The Hill exponent. It is functional response shape parameter.}
}
\value{
Returns a data frame with parameter values.
}
\description{
TBA
}
