% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_fit.R
\name{fr_fit}
\alias{fr_fit}
\title{Fitting of the functional response model with simulation by lsoda()}
\usage{
fr_fit(
  Nstart,
  Neaten,
  P,
  tend,
  tsteps = 100,
  lhs_iter_init = 1000,
  lhs_iter_val = 50,
  MC = T,
  noC = ceiling(parallel::detectCores()/2),
  range_mult = c(1.01, 1.1, 1.5, 2, 5),
  witer_max = 25,
  val_tol = 4
)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{tend}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{MC}{Should the code run in parallel? Default is TRUE.}

\item{noC}{The number of cores that should be used, default is 50% of all cores.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}

\item{lhs_iter}{Number of random latin hypercube samplings, default is 1000.}
}
\value{
Returns a vector of negative log-likelihoods in the order of
    Neaten/Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{
data(dfr_vp)

fr_fit(
  Neaten = dfr_vp$Neaten,
  Nstart = dfr_vp$Nstart,
  P = rep(1, length(dfr_vp$Neaten)),
  tend = rep(1, length(dfr_vp$Neaten))
)

}
