% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_fit.R
\name{fr_fit}
\alias{fr_fit}
\title{Fitting of the functional response model with simulation by lsoda()}
\usage{
fr_fit(
  Nstart,
  Neaten,
  P,
  tend,
  tsteps = 50,
  lhs_iter_init = 1280,
  lhs_iter_val = 320,
  MC = T,
  noC = ceiling(parallel::detectCores()/2),
  range_mult = c(1.001, 1.1, 2, 5),
  witer_max = 25,
  mle2_tol = 1e-12,
  val_tol = 6
)
}
\arguments{
\item{Nstart}{The initial prey/resource density.}

\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{P}{The consumer/predator abundance/counts/density.
Must be a vector of the length of Nstart.}

\item{tend}{The end time over up to which should be simulated (e.g. 24h).
Must be a vector of the length of Nstart.}

\item{tsteps}{The number of steps that should be returned by the solver.}

\item{lhs_iter_init}{Number of initial random latin hypercube samplings.}

\item{lhs_iter_val}{Number of validation random latin hypercube samplings.}

\item{MC}{Use parallel computing? (F/T)}

\item{noC}{If MC = T, how many cores to use?}

\item{range_mult}{The multipliers with which the current best parameters
should be multiplied for the validation random latin hypercube sampling.}

\item{witer_max}{How many fits should be performed without convergence?}

\item{mle2_tol}{The tolerance of a single mle2 fit.}

\item{val_tol}{The tolerance of the validation.}
}
\value{
TBA
}
\description{

}
\details{
TBA
}
\examples{

data(dfr_vp)

fr_fit(
  Neaten = dfr_vp$Neaten,
  Nstart = dfr_vp$Nstart,
  P = rep(1, length(dfr_vp$Neaten)),
  tend = rep(1, length(dfr_vp$Neaten))
)

}
