% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_nll.R
\name{fr_nll}
\alias{fr_nll}
\title{Calculation of negative log-likelihoods using lsoda() as method}
\usage{
fr_nll(Neaten, Nstart, P, tend, Fmax, N0, q, tsteps = 100)
}
\arguments{
\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{Nstart}{The initial prey/resource density.}

\item{tend}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{parms}{The model parameters as needed by e.g. \link[deSolve]{lsoda}.}
}
\value{
Returns a vector of negative log-likelihoods in the order of
    Neaten/Nstart.
}
\description{

}
\details{
The model parameters included in `parms` are (1) the maximum feeding
    rate `Fmax`, the half saturation density, `N0`, the shape parameter `q`,
    and the number or density of the predators/consumers, `P`.
}
\examples{

Nstart <- rep(1:100,10)
Neaten <- rbinom(length(Nstart),
                 size = Nstart,
                 prob = .5
                 )

fr_nll(
  Neaten = Neaten,
  Nstart = Nstart,
  P = rep(1, length(Nstart)),
  tend = rep(24, length(Nstart)),
  Fmax = 2,
  N0 = 5,
  q = 1
)


}
