% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs_parms.R
\name{lhs_parms}
\alias{lhs_parms}
\title{Random sampling of Parameters and Selection based on lhs}
\usage{
lhs_parms(
  Neaten,
  Nstart,
  P,
  tend,
  Fmax_range,
  N0_range,
  q_range,
  tsteps = 100,
  lhs_iter = 1000,
  MC = T,
  noC = ceiling(parallel::detectCores())
)
}
\arguments{
\item{Neaten}{The prey/resource items eaten throughout the experimental
trial.}

\item{Nstart}{The initial prey/resource density.}

\item{P}{TBA}

\item{tend}{The time over which should be simulated.}

\item{tsteps}{The number of times steps that \link[deSolve]{lsoda} should
return.}

\item{lhs_iter}{Number of random latin hypercube samplings, default is 1000.}

\item{MC}{Should the code run in parallel? Default is TRUE.}

\item{noC}{The number of cores that should be used, default is 50% of all cores.}

\item{Fmax}{TBA}

\item{N0}{TBA}

\item{q}{TBA}
}
\value{
Returns a data frame with parameter values.
}
\description{
TBA
}
